move mongo objects classes and put methods static
