+++ app/src/main/java/com/epsi/covidmanager/controller/DetailsSlot.java
@@ -5,7 +5,7 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 import com.epsi.covidmanager.R;
-import com.epsi.covidmanager.model.beans.Slot;
+import com.epsi.covidmanager.model.mongo.Slot;
 
 public class DetailsSlot extends AppCompatActivity {
